Use specific deps, add dev tls cert and key
Previously, package.json listed dependencies like "0.x" - now these are
more specific because after brushing the dust off this project and
running an `npm install && npm test`, shit was broken.

This latest version of mockgoose seems to be working now with the
specific verion of mongoose (which isn't the latest).