Store users in the db instead of conf.json
With this change, we can associate docs or anything else to specific
users (this isn't implemented yet) and modifying users won't require a
server restart.

Mainly changed the mkauth.js tool and the secure.js middleware for this
feature and added tests for both of them.

Also disabled mockgoose because it still isn't working properly.