Change docs API, update tests; still no UI
Deprecated the res-handler middleware - there's a long note in the file
explaining why. From a user's point of view - the API no longer returns
status, method, and errors properties in the json response.

Updated and added tests for docs - these include tests for user-doc
association.

The 'content' field on docs has been renamed to 'body'.

For the User model, 'username' has been renamed to 'name'.

Bumped up the package.json version to 1.0.0 but I won't publish it to
npm until there's an HTML UI for docs.

Docs API is now mounted at /api/docs, it ignores the Content-Type and
Accepts headers of the request. It will always try to parse json from
the request body.

Drop mongoose-unique-validator because we need to handle the Mongo 11000
error anyway on compound uniqueness among doc name and user.

Updated the index.html and docs.html pages and their css.